requirements.txt
# 語音處理相關
sounddevice>=0.4.6
pyaudio>=0.2.13
whisper>=1.1.10
pyttsx3>=2.90
gTTS>=2.3.2

# AI 與自然語言處理
google-cloud-aiplatform>=1.25.0
python-dateutil>=2.8.2
regex>=2023.5.5

# GUI 相關
tkinter
pillow>=10.0.0

# 資料處理
pandas>=2.0.0
numpy>=1.24.0
json5>=0.9.14
